+++ test/integration/translation_failures_test.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/kong/go-kong/kong"
 	"github.com/kong/kubernetes-testing-framework/pkg/clusters"
+	ktfkong "github.com/kong/kubernetes-testing-framework/pkg/clusters/addons/kong"
 	"github.com/kong/kubernetes-testing-framework/pkg/utils/kubernetes/generators"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -188,6 +189,48 @@ func TestTranslationFailures(t *testing.T) {
 				return expectedTranslationFailure{
 					causingObjects: []client.Object{ingress, service},
 					reasonContains: "can't find port for backend kubernetes service: no suitable port found",
+				}
+			},
+		},
+		{
+			name: "more than one certificate ref specified for a gateway listener",
+			translationFailureTrigger: func(t *testing.T, cleaner *clusters.Cleaner, ns string) expectedTranslationFailure {
+				secret1 := &corev1.Secret{ObjectMeta: metav1.ObjectMeta{
+					Name: testutils.RandomName(testTranslationFailuresObjectsPrefix),
+				}}
+				secret2 := &corev1.Secret{ObjectMeta: metav1.ObjectMeta{
+					Name: testutils.RandomName(testTranslationFailuresObjectsPrefix),
+				}}
+				secret1, err := env.Cluster().Client().CoreV1().Secrets(ns).Create(ctx, secret1, metav1.CreateOptions{})
+				require.NoError(t, err)
+				cleaner.Add(secret1)
+				secret2, err = env.Cluster().Client().CoreV1().Secrets(ns).Create(ctx, secret2, metav1.CreateOptions{})
+				require.NoError(t, err)
+				cleaner.Add(secret2)
+
+				gateway := deployGatewayReferringSecrets(t, cleaner, ns, secret1, secret2)
+
+				return expectedTranslationFailure{
+					causingObjects: []client.Object{gateway},
+					reasonContains: "more than one certificateRef",
+				}
+			},
+		},
+		{
+			name: "invalid secret referred by a gateway listener",
+			translationFailureTrigger: func(t *testing.T, cleaner *clusters.Cleaner, ns string) expectedTranslationFailure {
+				emptySecret := &corev1.Secret{ObjectMeta: metav1.ObjectMeta{
+					Name: testutils.RandomName(testTranslationFailuresObjectsPrefix),
+				}}
+				emptySecret, err := env.Cluster().Client().CoreV1().Secrets(ns).Create(ctx, emptySecret, metav1.CreateOptions{})
+				require.NoError(t, err)
+				cleaner.Add(emptySecret)
+
+				gateway := deployGatewayReferringSecrets(t, cleaner, ns, emptySecret)
+
+				return expectedTranslationFailure{
+					causingObjects: []client.Object{gateway, emptySecret},
+					reasonContains: "failed to construct certificate from secret",
 				}
 			},
 		},
@@ -449,4 +492,39 @@ func validService() *corev1.Service {
 			},
 		},
 	}
+}
+
+func deployGatewayReferringSecrets(t *testing.T, cleaner *clusters.Cleaner, ns string, secrets ...*corev1.Secret) *gatewayv1beta1.Gateway {
+	gatewayClient, err := gatewayclient.NewForConfig(env.Cluster().Config())
+	require.NoError(t, err)
+
+	gatewayClassName := testutils.RandomName(testTranslationFailuresObjectsPrefix)
+	gwc, err := DeployGatewayClass(ctx, gatewayClient, gatewayClassName)
+	require.NoError(t, err)
+	cleaner.Add(gwc)
+
+	certificateRefs := make([]gatewayv1beta1.SecretObjectReference, 0, len(secrets))
+	for _, s := range secrets {
+		sn := gatewayv1beta1.Namespace(s.GetNamespace())
+		certificateRefs = append(certificateRefs, gatewayv1beta1.SecretObjectReference{
+			Name:      gatewayv1beta1.ObjectName(s.GetName()),
+			Namespace: &sn,
+		})
+	}
+
+	gatewayName := testutils.RandomName(testTranslationFailuresObjectsPrefix)
+	hostname := gatewayv1beta1.Hostname(tlsRouteHostname)
+	gateway, err := DeployGateway(ctx, gatewayClient, ns, gatewayClassName, func(gw *gatewayv1beta1.Gateway) {
+		gw.Name = gatewayName
+		gw.Spec.Listeners = []gatewayv1beta1.Listener{{
+			Name:     gatewayv1beta1.SectionName(testutils.RandomName(testTranslationFailuresObjectsPrefix)),
+			Protocol: gatewayv1beta1.TLSProtocolType,
+			Port:     gatewayv1beta1.PortNumber(ktfkong.DefaultTLSServicePort),
+			Hostname: &hostname,
+			TLS:      &gatewayv1beta1.GatewayTLSConfig{CertificateRefs: certificateRefs},
+		}}
+	})
+	require.NoError(t, err)
+	cleaner.Add(gateway)
+	return gateway
 }